fix: use Long latency telemetry in Ultra dashboard

+++ app/src/main/java/com/cardenaspiero255/gamehubultra/UltraDashboard.kt
@@ -108,7 +108,7 @@ private fun UltraHeader() {
 private fun MetricsRow(
     device: DeviceInfo,
     refresh: Float?,
-    latency: Int?,
+    latency: Long?,
     battery: Int?,
     storage: Int?
 ) {
